refactor: popover props and more story
