+++ packages/components/popover/src/index.ts
@@ -2,72 +2,63 @@ import type { } from '@floating-ui/vue'
 
 export {
   OkuPopover,
-  createPopoverScope,
-  popoverProps,
-} from './popover'
-
-export type {
-  PopoverEmits,
-  PopoverProps,
 } from './popover'
 
 export {
   OkuPopoverAnchor,
-  popoverAnchorProps,
-} from './popoverAnchor'
-
-export type {
-  PopoverAnchorElement,
-  PopoverAnchorNaviteElement,
-  PopoverAnchorProps,
 } from './popoverAnchor'
 
 export {
   OkuPopoverTrigger,
-  popoverTriggerProps,
-} from './popoverTrigger'
-
-export type {
-  PopoverTriggerNaviteElement,
-  PopoverTriggerEmits,
-  PopoverTriggerProps,
 } from './popoverTrigger'
 
 export {
   OkuPopoverPortal,
-  popoverPortalProps,
-} from './popoverPortal'
-
-export type {
-  PopoverPortalProps,
 } from './popoverPortal'
 
 export {
   OkuPopoverContent,
-  popoverContentProps,
-} from './popoverContent'
-
-export type {
-  PopoverContentProps,
 } from './popoverContent'
 
 export {
   OkuPopoverClose,
 } from './popoverClose'
 
-export type {
-  PopoverCloseNaviteElement,
-  PopoverCloseElement,
-  PopoverCloseEmits,
-} from './popoverClose'
-
 export {
   OkuPopoverArrow,
-  popoverArrowProps,
 } from './popoverArrow'
 
 export type {
   PopoverArrowProps,
   PopoverArrowElement,
   PopoverArrowNaviteElement,
-} from './popoverArrow'
+
+  PopoverCloseNaviteElement,
+  PopoverCloseElement,
+  PopoverCloseEmits,
+
+  PopoverPortalProps,
+
+  PopoverTriggerNaviteElement,
+  PopoverTriggerEmits,
+  PopoverTriggerProps,
+
+  PopoverContentProps,
+
+  PopoverAnchorElement,
+  PopoverAnchorNaviteElement,
+  PopoverAnchorProps,
+
+  PopoverEmits,
+  PopoverProps,
+} from './props'
+
+export {
+  popoverArrowProps,
+  popoverTriggerProps,
+  popoverContentProps,
+  popoverAnchorProps,
+  createPopoverScope,
+  popoverProps,
+  popoverPortalProps,
+} from './props'
