+++ packages/components/popover/src/popoverContent.ts
@@ -1,46 +1,15 @@
-import type { PropType } from 'vue'
 import { computed, defineComponent, h, mergeProps, reactive, toRefs } from 'vue'
-import { primitiveProps } from '@oku-ui/primitive'
 import { reactiveOmit, useForwardRef } from '@oku-ui/use-composable'
 import { OkuPresence } from '@oku-ui/presence'
-import { scopePopoverProps } from './utils'
-import { usePortalInject } from './popoverPortal'
-import { usePopoverInject } from './popover'
-import type { PopoverContentTypeProps } from './popoverContentModal'
-import { OkuPopoverContentModal, popoverContentTypeProps } from './popoverContentModal'
+import { CONTENT_NAME, popoverContentProps, scopePopoverProps, usePopoverInject, usePortalInject } from './props'
+import { OkuPopoverContentModal } from './popoverContentModal'
 import { OkuPopoverContentNonModal } from './popoverContentNonModal'
-
-export const CONTENT_NAME = 'OkuPopoverContent'
-
-// TODO: PopoverContentTypeProps
-
-export interface PopoverContentProps extends PopoverContentTypeProps {
-  /**
-   * Used to force mounting when more control is needed. Useful when
-   * controlling animation with React animation libraries.
-   */
-  forceMount?: true
-}
-
-export const popoverContentProps = {
-  props: {
-    forceMount: {
-      type: Boolean as PropType<true | undefined>,
-      default: undefined,
-    },
-    ...popoverContentTypeProps.props,
-  },
-  emits: {
-    ...popoverContentTypeProps.emits,
-  },
-}
 
 const popoverContent = defineComponent({
   name: CONTENT_NAME,
   inheritAttrs: false,
   props: {
     ...popoverContentProps.props,
-    ...primitiveProps,
     ...scopePopoverProps,
   },
   emits: popoverContentProps.emits,
