+++ packages/components/popover/src/popoverContentNonModal.ts
@@ -1,23 +1,15 @@
 import { defineComponent, h, mergeProps, ref } from 'vue'
-import { primitiveProps } from '@oku-ui/primitive'
 import { useForwardRef } from '@oku-ui/use-composable'
-import { OkuSlot } from '@oku-ui/slot'
 import { composeEventHandlers } from '@oku-ui/utils'
 import { OkuPopoverContentImpl } from './popoverContentImpl'
-import { scopePopoverProps } from './utils'
-import { usePopoverInject } from './popover'
-import { CONTENT_NAME } from './popoverContent'
-import type { PopoverContentTypeEmits, PopoverContentTypeNaviteElement } from './popoverContentModal'
-import { popoverContentTypeProps } from './popoverContentModal'
-
-const NAME = 'OkuPopoverContentNonModal'
+import { CONTENT_NAME, CONTENT_NON_MODAL_NAME, popoverContentTypeProps, scopePopoverProps, usePopoverInject } from './props'
+import type { PopoverContentTypeEmits, PopoverContentTypeNaviteElement } from './props'
 
 const popoverContentNonModal = defineComponent({
-  name: NAME,
+  name: CONTENT_NON_MODAL_NAME,
   inheritAttrs: false,
   props: {
     ...popoverContentTypeProps.props,
-    ...primitiveProps,
     ...scopePopoverProps,
   },
   emits: popoverContentTypeProps.emits,
@@ -29,51 +21,49 @@ const popoverContentNonModal = defineComponent({
 
     const forwardedRef = useForwardRef()
 
-    return () => h(OkuSlot, {}, {
-      default: () => h(OkuPopoverContentImpl, {
-        ...mergeProps(attrs, props),
-        ref: forwardedRef,
-        trapFocus: false,
-        disableOutsidePointerEvents: false,
-        onCloseAutoFocus: composeEventHandlers<PopoverContentTypeEmits['closeAutoFocus'][0]>((event) => {
-          emit('closeAutoFocus', event)
+    return () => h(OkuPopoverContentImpl, {
+      ...mergeProps(attrs, props),
+      ref: forwardedRef,
+      trapFocus: false,
+      disableOutsidePointerEvents: false,
+      onCloseAutoFocus: composeEventHandlers<PopoverContentTypeEmits['closeAutoFocus'][0]>((event) => {
+        emit('closeAutoFocus', event)
 
-          if (!event.defaultPrevented) {
-            if (!hasInteractedOutsideRef.value)
-              inject.triggerRef.value?.focus()
+        if (!event.defaultPrevented) {
+          if (!hasInteractedOutsideRef.value)
+            inject.triggerRef.value?.focus()
             // Always prevent auto focus because we either focus manually or want user agent focus
-            event.preventDefault()
-          }
+          event.preventDefault()
+        }
 
-          hasInteractedOutsideRef.value = false
-          hasPointerDownOutsideRef.value = false
-        }),
-        onFocusoutSide: composeEventHandlers<PopoverContentTypeEmits['focusoutSide'][0]>((event) => {
-          emit('focusoutSide', event)
-          if (!event.defaultPrevented) {
-            hasInteractedOutsideRef.value = true
-            if (event.detail.originalEvent.type === 'pointerdown')
-              hasPointerDownOutsideRef.value = true
-          }
-
-          // Prevent dismissing when clicking the trigger.
-          // As the trigger is already setup to close, without doing so would
-          // cause it to close and immediately open.
-          const target = event.target as HTMLElement
-          const targetIsTrigger = inject.triggerRef.value?.contains(target)
-          if (targetIsTrigger)
-            event.preventDefault()
-
-          // On Safari if the trigger is inside a container with tabIndex={0}, when clicked
-          // we will get the pointer down outside event on the trigger, but then a subsequent
-          // focus outside event on the container, we ignore any focus outside event when we've
-          // already had a pointer down outside event.
-          if (event.detail.originalEvent.type === 'focusin' && hasPointerDownOutsideRef.value)
-            event.preventDefault()
-        }),
-      }, {
-        default: () => slots.default?.(),
+        hasInteractedOutsideRef.value = false
+        hasPointerDownOutsideRef.value = false
       }),
+      onFocusoutSide: composeEventHandlers<PopoverContentTypeEmits['focusoutSide'][0]>((event) => {
+        emit('focusoutSide', event)
+        if (!event.defaultPrevented) {
+          hasInteractedOutsideRef.value = true
+          if (event.detail.originalEvent.type === 'pointerdown')
+            hasPointerDownOutsideRef.value = true
+        }
+
+        // Prevent dismissing when clicking the trigger.
+        // As the trigger is already setup to close, without doing so would
+        // cause it to close and immediately open.
+        const target = event.target as HTMLElement
+        const targetIsTrigger = inject.triggerRef.value?.contains(target)
+        if (targetIsTrigger)
+          event.preventDefault()
+
+        // On Safari if the trigger is inside a container with tabIndex={0}, when clicked
+        // we will get the pointer down outside event on the trigger, but then a subsequent
+        // focus outside event on the container, we ignore any focus outside event when we've
+        // already had a pointer down outside event.
+        if (event.detail.originalEvent.type === 'focusin' && hasPointerDownOutsideRef.value)
+          event.preventDefault()
+      }),
+    }, {
+      default: () => slots.default?.(),
     })
   },
 })
